# Reorder inputs for Distance and Intersects